# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ademir API",
    "description": "API usada pelo novo frontend (React/Vite). Endpoints sob o prefixo /api.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3101/api" }
  ],
...
  "paths": {
...
    "/bilheteria/config": {
      "get": { "tags": ["bilheteria"], "summary": "Obter config da bilheteria", "responses": { "200": { "description": "Config" } } },
      "post": {
        "tags": ["bilheteria"],
        "summary": "Salvar config",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Config salva" } }
      }
    },
    "/bilheteria/processar-pagamento": {
      "post": {
        "tags": ["bilheteria"],
        "summary": "Processar pagamento",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Resultado do pagamento" } }
      }
    },
    "/bilheteria/vender": {
      "post": {
        "tags": ["bilheteria"],
        "summary": "Vender ingressos",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "tipo": { "type": "string" }, "quantidade": { "type": "integer" } } } } } },
        "responses": { "200": { "description": "Ingressos gerados" } }
      }
    },
...
    "/bilheteria/liberar": {
      "post": {
        "tags": ["bilheteria"],
        "summary": "Liberar ingresso",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "codigo": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Ingresso liberado" }, "404": { "description": "Não encontrado" } }
      }
    },
...
  }
}